Add files stock to user management and user details
Many of these files will need to be modified at some point.
For instance, Spring-Security is still in an xml file instead of in the
java configuration. For help with this migration, reference:
http://docs.spring.io/spring-security/site/migrate/current/3-to-4/html5/migrate-3-to-4-xml.html.

Also, the user tracker may not be reliable considering I couldn't get it
to work with this build. It requires HttpSessionListener to have the
@Service annotation which is commented out for now.

Otherwise this build will at least start running.