Topic Approval System Overhaul - PHP Inital Commit
Moving from Java to PHP.
Even though there are plenty of perks to using this framework, (i.e. Security, speed) setup is not necessarily easy.
For the sake of faster deployment considering the primary client will be using a simple Apache server and PHP Server, PHP is a frendly alternative to a Java Application Servlet Container.