Starting Course Management
Course Management is 'working' but has unveiled a fatal design flaw.
It would seem a user's authorities pertaining to product use require a separate realm of handling. In this case, a user should either be an admin, or not one.
This would leave the other values in the Role table in an incorrect place (stock functionality would lead us to assume those values are only 'ROLE_PROFESSOR', 'ROLE_TA', & 'ROLE_STUDENT').

The solution for this should be to add a UserAuthority and Authority table for mapping a user's unique product handling necessities to Authorities rather than through a Role.
The Role and UserRole table can in fact stay, however UserRole will now have foreign keys identifying it not to a User, but an entry in the UserCourse table; thereby enforcing the concept that an individual User can be enrolled in multiple courses with a different Role in each of them but of course, only serving ONE role per course.
I.E. a student of 756 can be a STUDENT, but also a TA of 646 where both courses require students to submit topics as part of an assignment.

*** NOTE: Be mindful of the language currently used to identify a user role, as the naming convention needs to be adjusted in the next version to STRICTLY DISTINGUISH between a UserAuthority and a UserRole!

Outside of this design issue, user management (login, logout, sessions, profile editing) works perfectly and securely never giving too much power to non-administrators.