Data Layer Fixes & Course Management
Hopefully, this is the final time I need to alter the schema.
Topic now has username and course number for whoever submitted the topic and for what course.
Otherwise, it would have been to confusing for users.

Course Management is underway. The JSON endpoints are working great and the widgets are getting there.
Only need to tweak the user interface for updating a course (name & roster) & adding a new course.

Then onto topic management, followed by the interface for a user topic submission!

(If there is time, the features I want to pitch are mass user additions and auto-topic-approval!)